refactor: Split pipeline into modular components
- Add pipeline_nodes.py with node functions (stt_node, should_retry_*)
- Add pipeline_agents.py with agent functions (rewrite_*, evaluation_*, etc.)
- Add pipeline_builder.py with pipeline graph construction
- Simplify graph_pipeline.py to only contain execution functions
- Improve code organization and maintainability
- Reduce file sizes from 800+ lines to manageable chunks
